Pumpkin SMASH
A 2023 Halloween whack-a-mole type game but with pumpkins.

The pumpkins spawn in one of the 28 spawn point positions, there can be only one pumpkin per spawn point. The pumpkin can either be destroyed by a player or it de-spawns after some time.

The player controls a hammer by clicking. The hammer has an animation made inside unity and when the hammer is all the way down the collider gets enabled, otherwise it is disabled so that the player can't hit multiple pumpkins with one swing - at most they can hit 2 if they spawn besides each other.

I've added some sounds, fog, and a bunch of synty studio models.

There are 2 texts, one for displaying smashed pumpkin counter and the other showing how many pumpkins we can still miss before it is game over. On game over and at start there is a main menu with instructions and a start button.

The pumpkin spawner code was mostly made with chat gpt.